v0.3.1140 — SCALE-SEAM ㉚, GMP and pay-app stack into cost.ts
Eight methods (costSummary, gmpBudget, baseline/variance/cashflow, SOV, G702, pay-app
invoice) move onto the existing withCost mixin. pxSummary and costTraceability stay.
client.ts 2641 → 2580. MARKS widened to ㉛.
