+++ apps/web/src/api/cost.ts
@@ -1,24 +1,15 @@
-/** Cost reconciliation across documents and cost codes: margin, cost-spine, commercial drift.
+/** Cost reconciliation, vintages, and the GC GMP / pay-app stack.
  *
- *  Route-group `/projects/{pid}/{margin,cost-spine,commercial-drift}`, taken out of `client.ts`
- *  because the size ratchet in `services/api/test_file_sizes.py` requires it — the file sat exactly
- *  on its 3,780 pin, and "a new endpoint added straight to `client.ts` will fail this, and that
- *  friction is the point". So the group moved and the pin moved DOWN.
+ *  SCALE-SEAM ㉚. Grouped by what they ANSWER: *what does the GC job cost, and how is the GMP billed?*
+ *  `costSummary` sat ~140 lines above `gmpBudget` in `client.ts`, with versions, safety, QTO and
+ *  smart-views in between. `pxSummary` sat IN the GMP run and did **not** come — it answers whether
+ *  the job is on track (schedule + budget together), not what the GMP is. `costTraceability` stayed
+ *  too: ㉘ already recorded it as a takeoff question. *Adjacency in a file is not a relationship.*
  *
- *  **`commercialDrift` is new, and it exists because the endpoint was unreachable** —
- *  `test_route_reachability` flagged it as a new uncalled route, correctly, and the honest fix is a
- *  caller rather than an exemption.
- *
- *  The three are one family on purpose. `marginByCostCode` and `costSpine` both measure **per cost
- *  code**; `commercialDrift` measures **per document**. A roll-up adds before it compares, so two
- *  subcontracts can net to the right code total while one award drifted up and another down — which
- *  is invisible to the first two and is exactly what the third is for. Anyone reading one of these
- *  should know the other axis exists.
- *
- *  A mixin, so every call site resolves unchanged.
+ *  Composed through the existing `withCost` wrapper — no extra `withX()` on `ApiClient`.
  */
 import { HttpCore } from "./httpCore";
-import type { ResolveAction } from "./types";
+import type { ModuleRecord, ResolveAction } from "./types";
 
 type Ctor<T> = new (...args: any[]) => T;
 
@@ -117,6 +108,67 @@ export function withCost<TBase extends Ctor<HttpCore>>(Base: TBase) {
       code_count: number; usable_count: number; min_projects: number;
       message: string | null;
     }>(`/benchmarks/unit-rates?min_projects=${minProjects}`);
+  }
+
+  /** costSummary — job-cost rollup: budget, committed, actual, forecast, projected over/under. */
+  costSummary(pid: string) {
+    return this.json<{ budget: number; committed: number; actual: number; forecast: number; projected_over_under: number; pct_committed: number; pct_spent: number }>(
+      `/projects/${pid}/cost/summary`);
+  }
+  /** Full GC project budget (GMP): direct + GC/GR + overhead/fee/contingency, each budget vs
+   *  committed vs actual vs variance; reconciled to the prime contract + developer proforma. */
+  gmpBudget(pid: string) {
+    type Cat = { key: string; name: string; budget: number; committed: number; actual: number;
+      forecast: number; eac: number; etc: number; variance: number; lines: { name: string; budget: number;
+      committed: number; eac?: number; etc?: number; variance: number; is_group?: boolean }[];
+      groups?: { name: string; budget: number }[] };
+    return this.json<{
+      gmp: { contract_value: number; computed: number; reconciliation: number | null; cost_of_work: number;
+        approved_changes?: number; unallocated_changes?: number; revised?: number;
+        markups: { overhead_pct: number; fee_pct: number; contingency_pct: number } };
+      categories: Cat[];
+      totals: { budget: number; committed: number; actual: number; forecast: number; eac: number; etc: number; variance: number };
+      completion: { bac: number; eac: number; etc: number; actual_to_date: number; projected_over_under: number; pct_spent: number };
+      bid_packages: { ref: string; name: string; trade?: string; budget: number; awarded: number;
+        bought_out: boolean; savings: number; submissions: number }[];
+      buyout: { packages: number; bought_out: number; budget: number; awarded: number; savings: number };
+      staffing: { projected: number; headcount_roles: number };
+      proforma: { hard_cost: number; gmp_vs_hard: number } | null;
+    }>(`/projects/${pid}/budget/gmp`);
+  }
+  /** Snapshot the current GMP budget as the baseline (for budget-movement tracking). */
+  setBudgetBaseline(pid: string) {
+    return this.json<{ captured_at: string; gmp_computed: number; lines: number }>(
+      `/projects/${pid}/budget/baseline`, { method: "POST" });
+  }
+  /** Budget movement vs the baseline (per category + line). Rejects if no baseline set. */
+  budgetVariance(pid: string) {
+    return this.json<{ captured_at: string; baseline_gmp: number; current_gmp: number; total_delta: number;
+      categories: { key: string; baseline: number; current: number; delta: number }[];
+      lines: { code: string; baseline: number; current: number; delta: number }[] }>(
+      `/projects/${pid}/budget/variance`);
+  }
+  /** Cost-loaded schedule → monthly cash-flow / draw curve (construction S-curve). */
+  budgetCashflow(pid: string) {
+    return this.json<{ total: number; months: number; loaded_activities: number; peak_month_cost: number;
+      series: { month: string; cost: number; cumulative: number; pct: number }[] }>(
+      `/projects/${pid}/budget/cashflow`);
+  }
+  /** Seed the owner pay-app SOV from the GMP budget lines (idempotent unless replace). */
+  sovFromBudget(pid: string, replace = false) {
+    return this.json<{ created: number; lines?: number; scheduled_value?: number; skipped?: number; note?: string }>(
+      `/projects/${pid}/cost/sov/from-budget?replace=${replace}`, { method: "POST" });
+  }
+  /** The owner pay application (G702 certificate + G703 continuation) as a signable PDF blob. */
+  async payAppPdf(pid: string, appNo = 1) {
+    const res = await fetch(this.url(`/projects/${pid}/cost/g702.pdf?app_no=${appNo}`), { headers: this.authHeaders() });
+    if (!res.ok) throw new Error(`pay-app PDF -> ${res.status}`);
+    return res.blob();
+  }
+  /** Create an owner-invoice record from the current pay application (amount = current payment due). */
+  payAppInvoice(pid: string, appNo = 1) {
+    return this.json<{ owner_invoice: ModuleRecord; application_no: number; amount: number }>(
+      `/projects/${pid}/cost/pay-app/invoice`, { method: "POST", body: JSON.stringify({ app_no: appNo }) });
   }
   };
 }
